requirements/development.txt
-r common.txt

celery==3.1.7
django-debug-toolbar==1.2.1
coverage==3.7.1
nose==1.3.3
django-nose==1.2
